ssrw_requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pyyaml
kornia>=0.6.0
lpips
tqdm